refactor samplesheet prep service to helper
